+++ server/models.py
@@ -108,20 +108,6 @@ class LaundrySnapshot(sqldb.Model):
     total_dryers = sqldb.Column(sqldb.Integer, nullable=False)
 
 
-class StudySpacesBooking(sqldb.Model):
-    id = sqldb.Column(sqldb.Integer, primary_key=True)
-    account = sqldb.Column(sqldb.VARCHAR(255), sqldb.ForeignKey('account.id'), nullable=True)
-    user = sqldb.Column(sqldb.Integer, sqldb.ForeignKey('user.id'), nullable=True)
-    booking_id = sqldb.Column(sqldb.Text, nullable=True)
-    date = sqldb.Column(sqldb.DateTime, default=datetime.datetime.now)
-    lid = sqldb.Column(sqldb.Integer, nullable=True)
-    rid = sqldb.Column(sqldb.Integer, nullable=True)
-    email = sqldb.Column(sqldb.Text, nullable=True)
-    start = sqldb.Column(sqldb.DateTime, nullable=True)
-    end = sqldb.Column(sqldb.DateTime, nullable=True)
-    is_cancelled = sqldb.Column(sqldb.Boolean, default=False)
-
-
 class User(sqldb.Model):
     id = sqldb.Column(sqldb.Integer, primary_key=True)
     created_at = sqldb.Column(sqldb.DateTime, server_default=sqldb.func.now())
